+++ src/Api/Connector/AuthInit.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Smolblog\RestApiBase\Connector;
+namespace Smolblog\Api\Connector;
 
 use Smolblog\Framework\Objects\Identifier;
 use Smolblog\Framework\Objects\Value;
-use Smolblog\RestApiBase\Endpoint;
-use Smolblog\RestApiBase\EndpointConfig;
-use Smolblog\RestApiBase\Exceptions\NotFound;
-use Smolblog\RestApiBase\GenericResponse;
-use Smolblog\RestApiBase\ParameterType;
+use Smolblog\Api\Endpoint;
+use Smolblog\Api\EndpointConfig;
+use Smolblog\Api\Exceptions\NotFound;
+use Smolblog\Api\GenericResponse;
+use Smolblog\Api\ParameterType;
 
 /**
  * Kick off an OAuth request to an external provider.
